Move files into folders
Execute move
