--- EquipListener.java
@@ -1,6 +1,0 @@
-package SpriteAnimator.Listeners;
-import SpriteAnimator.Events.EquipEvent;;
-
-public interface EquipListener {
-	void eventReceived(EquipEvent arg0);
-}
