+++ Listeners/EquipListener.java
@@ -1,0 +1,6 @@
+package SpriteAnimator.Listeners;
+import SpriteAnimator.Events.EquipEvent;;
+
+public interface EquipListener {
+	void eventReceived(EquipEvent arg0);
+}
