+++ Listeners/ModeListener.java
@@ -1,0 +1,7 @@
+package SpriteAnimator.Listeners;
+
+import SpriteAnimator.Events.ModeEvent;
+
+public interface ModeListener {
+	void eventReceived(ModeEvent arg0);
+}
