+++ Listeners/SpeedListener.java
@@ -1,0 +1,7 @@
+package SpriteAnimator.Listeners;
+
+import SpriteAnimator.Events.SpeedEvent;
+
+public interface SpeedListener {
+	void eventReceived(SpeedEvent arg0);
+}
