+++ Listeners/StepListener.java
@@ -1,0 +1,7 @@
+package SpriteAnimator.Listeners;
+
+import SpriteAnimator.Events.StepEvent;
+
+public interface StepListener {
+	void eventReceived(StepEvent arg0);
+}
