+++ Listeners/ZoomListener.java
@@ -1,0 +1,7 @@
+package SpriteAnimator.Listeners;
+
+import SpriteAnimator.Events.ZoomEvent;
+
+public interface ZoomListener {
+	void eventReceived(ZoomEvent arg0);
+}
