--- ModeListener.java
@@ -1,7 +1,0 @@
-package SpriteAnimator.Listeners;
-
-import SpriteAnimator.Events.ModeEvent;
-
-public interface ModeListener {
-	void eventReceived(ModeEvent arg0);
-}
