--- SpeedListener.java
@@ -1,7 +1,0 @@
-package SpriteAnimator.Listeners;
-
-import SpriteAnimator.Events.SpeedEvent;
-
-public interface SpeedListener {
-	void eventReceived(SpeedEvent arg0);
-}
