+++ Sprite/SpriteManipulator.java
