--- StepListener.java
@@ -1,7 +1,0 @@
-package SpriteAnimator.Listeners;
-
-import SpriteAnimator.Events.StepEvent;
-
-public interface StepListener {
-	void eventReceived(StepEvent arg0);
-}
