--- ZoomListener.java
@@ -1,7 +1,0 @@
-package SpriteAnimator.Listeners;
-
-import SpriteAnimator.Events.ZoomEvent;
-
-public interface ZoomListener {
-	void eventReceived(ZoomEvent arg0);
-}
